initial attempt at converting to ss4
